requirements.txt
coinbase-advanced-py
python-dotenv
pydantic>=2.0
pydantic-settings
structlog
prometheus-client
backoff
fastapi
uvicorn

# Test dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0
coverage-badge>=1.1.0
coverage>=7.0.0